Move exceptions to exceptions.py
We were running into the problem where there were individual Exceptions
in individual files, and reusing exceptions (for example if there was
to be a bulk API python file) would require importing from the
single-object API file.

Moving all the exceptions to this file would make the most sense.
