+++ simple_salesforce/tests/test_login.py
@@ -20,10 +20,8 @@ except ImportError:
     from urllib.parse import urlparse
 
 from simple_salesforce import tests
-from simple_salesforce.login import (
-    SalesforceLogin, SalesforceAuthenticationFailed
-)
-
+from simple_salesforce.login import SalesforceLogin
+from simple_salesforce.exceptions import SalesforceAuthenticationFailed
 
 
 class TestSalesforceLogin(unittest.TestCase):
