+++ simple_salesforce/util.py
@@ -34,23 +34,3 @@ def date_to_iso8601(date):
         timezone=timezone_str
         ).replace(':', '%3A').replace('+', '%2B')
 
-
-class SalesforceError(Exception):
-    """Base Salesforce API exception"""
-
-    message = u'Unknown error occurred for {url}. Response content: {content}'
-
-    def __init__(self, url, status, resource_name, content):
-        # TODO exceptions don't seem to be using parent constructors at all.
-        # this should be fixed.
-        # pylint: disable=super-init-not-called
-        self.url = url
-        self.status = status
-        self.resource_name = resource_name
-        self.content = content
-
-    def __str__(self):
-        return self.message.format(url=self.url, content=self.content)
-
-    def __unicode__(self):
-        return self.__str__()
